% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulationExperiment.R
\name{generate_from_simple_logistic_model}
\alias{generate_from_simple_logistic_model}
\title{Simulate data from a simple logistic model}
\usage{
generate_from_simple_logistic_model(n = 1000)
}
\description{
Simulate data from a simple logistic model
}

