% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Predict the y value for all data points,
given  and the sample from posterior of betas}
\usage{
get_predictions(X, beta)
}
\arguments{
\item{X}{A design matrix.}

\item{beta}{Sample from the posterior distribution of betas. Alternatively, could be a vector with estimated betas.}
}
\description{
Predict the y value for all data points,
given  and the sample from posterior of betas
}

