% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Gibbs two-step sampling procedure
for parameter vector beta and the latent Polya-Gamma variables}
\usage{
gibbs_sampler(y, X, lambda = 1e-04, b = rep(0, ncol(X)), B = lambda *
  diag(ncol(X)), n_iter_total = 200, burn_in = 100, naive = FALSE,
  naive_n_terms = 100, t = 0.64)
}
\arguments{
\item{y}{binary vector of observations}

\item{X}{design matrix of covariates}

\item{lambda}{The diagonal element of the precision matrix of the prior distribution (see also parameter B)}

\item{b}{The prior mean of the prior distribution. Defaults to a vector of zeros.}

\item{B}{The prior precision of the prior distribution. Defaults to lambda * identity.}

\item{naive}{Should the naive approximation be used to generate the Polya-Gamma distribution}

\item{naive_n_terms}{If the naive approximation is used, then this specifies number of terms in the finite sum.}

\item{t}{The parameter in the accept-reject algorithm for sampling from Polya-Gamma distribution (see paper for details).}

\item{n_iter}{The total number of iterations in the MCMC chain}
}
\value{
list containing the MCMC samples from the posterior distribution of beta
}
\description{
Gibbs two-step sampling procedure
for parameter vector beta and the latent Polya-Gamma variables
}
\examples{
data = generate_from_simple_logistic_model(n=100)
obj = gibbs_sampler(data$y, data$X, lambda=0.001, n_iter_total=100, burn_in=50)
plot(obj)
}

