% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{plot.PG}
\alias{plot.PG}
\title{Traceplot and autocorrelation plot for the beta chain}
\usage{
\method{plot}{PG}(obj, which_parameters = 1:ncol(obj$beta))
}
\arguments{
\item{which_parameters}{Vector of indexes which components of beta vector should be plotted.}
}
\description{
Traceplot and autocorrelation plot for the beta chain
}

