% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Gibbs two-step sampling procedure
for parameter vector beta and the latent Polya-Gamma variables}
\usage{
gibbs_sampler(y, X, lambda = 1e-04, b = rep(0, ncol(X)), B = lambda *
  diag(ncol(X)), n_iter = 100, naive = FALSE, naive_n_terms = 100,
  t = 0.64)
}
\arguments{
\item{y}{binary vector of observations}

\item{X}{design matrix of covariates}

\item{lambda}{}
}
\value{
list containing the MCMC samples from the posterior distribution of beta
}
\description{
Gibbs two-step sampling procedure
for parameter vector beta and the latent Polya-Gamma variables
}

